refactored PictureOfTheDayViewModelTests.swift to isolate api mocking
